feat: 🎸 Explicitly declare nodeClass and nodeType in core types
This is necessary to ensure disambiguation between block, inline, and
text elements. Additionally, this codifies the rich text specification
as provided by the CEP (alternatives wouldn't make it through our
validation layer, and should not be supported as such).

BREAKING CHANGE: Potentially breaks TypeScript libraries pulling this in as a dependency
if they are not explicitly stating nodeClass and nodeType.