fix: gracefully handle empty rich text documents
When a rich text field is empty in the contentful ui it is returned as
`undefined`. If passed into `documentToHtmlString` it would throw as `node.map`
isn't a function. Instead of throwing now it will simply return an empty string
if passed input that doesn't contain nodes or is `undefined`.