fix: 🐛 incorrect order of links generated
The addLinksFromRichTextNode crawls through the nodes and always picks
up the last item in that array. This causes links to be entity links to
be generated in the incorrect order. Instead of taking the last node,
the first one is picked up to parse the nodes in the order they appeared
in the rich text